Use overlap-discard on the LPF filter to improve its performance
This reduces the ringing noise up to 10dB depending on the song,
however in certain songs it's still audible. Note that on the air
where the RF noise dominates and the receivers use power squelch
filters that let only signals of a certain amplitute pass,
this distortion is rarely audible. On the other hand I don't like
seeing this on the spectrum analyzer nor hearing it localy.

Many thanks to Manolis Surligas @surligas for his help !