Switch to a sinc filter multiplied by a Nutall window
I tried with the Gausian to get away without complex representation
and multiplication of the filter's curve but its responce is too
steep and ringy. I even played with it trying to keep the curve
above zero etc wich gave ~1dB less but nothing great.

I decided to switch back to a windowed sinc which is the typical
way of designing FIR filters. The result is much better and now
that we use floats instead of doubles is not that bad cpu-wise.
This reduces the filter's noice even further, I now get a spike
near the cutoff frequency but its amplitude is way too low to
be audible.